restructure unittests
